Add new programs and templates for penaumtics
Add a template for safety shutter, wher ethe PSS permit signals are created.
Add a program for the Pneumatic_Box, where all the signals for pressure sensors
and pressure calcualtions are implemented.
Correct the Pneumatics_Template TaskInfo was missing a _, delete the pressurelimtis,
and add the pressure group selection.
Call the new programs from PROG() MAIN.